profile packageの中身をprofile.goに集約
